chore: M2 close — code bug fixes, doc audit, milestone status update
Code fixes:
- PhaseRunner.cs: int→long in foreach over EntityIds (silent truncation of entity IDs >2B in EventEntities table)
- CharacterBehaviorPhase.cs: fix operator precedence on & 0x7FFFFFFF in civ-born seq derivation (50_000 offset was excluded from mask, breaking reproducibility)

Docs updated:
- CLAUDE.md: milestone status M1→M2 complete, session workflow step 3, Spotlight milestone label
- interface_contracts.md: v0.4→v0.5, PendingEvent EntityIds field, EntitySnapshot CivName field, CivId.IsValid, EventEntities table schema + FK ordering note, EventType ranges (M2 beast/character/civ/population), IWorldStateReadOnly GetRelationship uncommented, Settlements/Civilizations accessors, future M3 stubs
- architecture_decision_records.md: v0.1→v0.2, ADRs 025-029 (CivId value type, PopulationF accumulator, CharacterNamesConfig, EventEntities junction table, civ-born character generation)
- mvp_spec.md: v0.1→v0.3, M2 marked COMPLETE with implementation summary per phase
- m2_phase3_tier2_characters.md, m2_phase4_tier3_population.md: Status IN PROGRESS→COMPLETE

+++ WorldEngine.Sim/Simulation/PhaseRunner.cs
@@ -158,7 +158,7 @@ public sealed class PhaseRunner
             if (pe.CauseEventId.HasValue && pe.CauseEventId.Value.Value != 0)
                 edges.Add((pe.CauseEventId.Value.Value, evId));
             if (pe.EntityIds is { Count: > 0 })
-                foreach (int eid in pe.EntityIds)
+                foreach (long eid in pe.EntityIds)
                     entPairs.Add((evId, eid));
         }
         if (edges.Count    > 0) _eventStore.InsertCausalEdges(edges);
